+++ include/SABER/ProgSlice.h
@@ -30,7 +30,7 @@
 #ifndef PROGSLICE_H_
 #define PROGSLICE_H_
 
-#include "Util/PathCondAllocator.h"
+#include "SABER/PathCondAllocator.h"
 #include "Util/WorkList.h"
 #include "Graphs/SVFG.h"
 #include "Util/DPItem.h"
